refractory spacing issues

+++ src/thermal/simulations/Numerical_Temperature.cpp
@@ -150,9 +150,9 @@ void temperature_1D(const class physicalModel::TBCsystem TBCsystem,
                     MatrixArrays->B2, MatrixArrays->B3, MatrixArrays->B4,
                     k1_thermal, k2_thermal, psi1_thermal, psi2_thermal );
 
-      solveMatrix( MatrixArrays->M2 , MatrixArrays->A1 , MatrixArrays->A2 ,
-                   MatrixArrays->A3 , MatrixArrays->b ,
-                   MatrixArrays->Temperature );
+      solveMatrix( static_cast<int>( MatrixArrays->M2 ) , MatrixArrays->A1 ,
+                   MatrixArrays->A2 , MatrixArrays->A3 , MatrixArrays->b ,
+                   MatrixArrays->Temperature ) ;
 
 
       for (size_t j = 0 ; j < mesh.M2 ; ++j )
